requirements.txt
mlflow>=2.5.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece
rouge-score
datasets
pandas
scikit-learn
pyyaml
python-dotenv